Fix "value" setter in datagrid component:
First, self.rows was not mutated by the setter like it is in
load_value.  This means many tests would fail because accessing
the rows in the datagrid would not return anything.

Second, the expected structure of "rows" is a list of dicts.  The
dicts should map the form key (a string) to the value (a Component
object).  Instead, the original code would have the rows as a list
of lists, which would contain 1 entry which would be a dict.

So, from the fixture example the first datagrid would get encoded as
follows:

[[{email: 'bob@example.com'}, {registrationDateTime: datetime(2021-04-05...)}],
 [{email: 'foo@example.com'}, {registrationDateTime: datetime(2021-04-06...)}]]

Instead, what it should be:

[{email: 'bob@example.com', registrationDateTime: datetime(2021-04-05...)},
 {email: 'foo@example.com', registrationDateTime: datetime(2021-04-06...)}]

Thirdly, the above is also not entirely correct; the value should be
wrapped in a component object; instead, the original code would fetch
the component and call _encode_value() on the value, but not actually
instantiate the component and set the value inside the component
object.