Introduce new concept of "component owner"
This is a bit like a combination between the parent component and a
builder; it's the object responsible for instantiating the component
in question.  So, normally this will be the builder, but inside a
datagrid, each component has the datagrid as the component owner.

If one were to nest datagrids, the inner datagrid would have the outer
datagrid as its component owner.

This is required so that the form_component keys match those of the
form JSON, where the datagrids are in the builder's form_components
but the children of the datagrids are not in the builder's
form_components (but they can be found in the datagrid's
form_components, which is a new property for this class).

This also slightly refactors the way the form_components are set
inside the builder (i.e., in the owner component); instead of doing
this straight inside the load() method, this now happens when setting
the component owner.  This is a bit iffy but matches the way parent
is set, which will also set the components in the parent, for example.

Another new property "is_form_component" is introduced which
simplifies the check for whether the component should be added to the
form_components dict.  This can simply be overridden in component
classes which shouldn't be considered form components (i.e., pure
rendering or layout components).

Finally, the child_component_owner property is a simple way to keep
load() the way it is and still have it pass a different component
object depending on subclass.  In dataGrid, the child_component_owner
is the datagrid itself.  In other components, the
child_component_owner is just its component_owner (passing it on
recursively down the tree until a dataGrid is hit).

The renderer is still a bit of a mess and needs some changes to make
it work proprly.  For now it simply proxies the load_components and
components property to its builder, this breaks the least number of
tests.