Major refactor of form value loading and datagrid representation
There was a certain confusion between components instantiated for a
builder and components instantiated for a form (which is then also
used for rendering).

The load function would be invoked without data in the builder, which
would build up a sort of "blueprint" object tree representing the
abstract form, and later the load function would be invoked on this
tree by the Form class _with_ data, which would essentially clone the
tree with instantiated copies inside a datagrid.  But because the form
data was keyed on items inside, the form data would always contain
only the data of the last row inside the grid.

To fix this, a new "GridRow" class is introduced which acts as a
component owner for each instance of the components inside the
datagrid, so that it establishes a new scope for each row to avoid
capturing the latest only.

The load function did a little too much, so it is split up into load,
which handles the setting of the parent and component owner and then
calling load_value which is the new method that actually implements
the instantiation of subcomponents and loading of values inside it.
This method is overridden in datagrid to create the aforementioned
GridRows.  The value setter in datagrid is no longer required, as
load_value does all the work of instantiating the child components.

Also rename load_value to load_data, as that is slightly better naming
for what it (now) does.