Implement 'row' data for JSON logic visibility in data grids
When evaluating JSON conditional visibility, data grids will have a
'row' property which provides access to the data of that row only.
Also, 'data' will still be available and have the entire form data.

While we're at it, for completeness in the tests we also need HTML
rendering for datagrids and column components because those are both
used in this test case's builder JSON.  This is a quick hack which
simply renders a table.  More complete support would be required to
render the column widths properly etc.