Fix bug that caused subcomponents to be double-processed (#17)
The code that's marked as "iffy" in load_data would be processing
input components nested under other components inside panels (or other
pure layout components), even if the loop above it had already
processed them through the child component, causing a misparenting and
double instantiation of input components to happen.

As a hack, we skip processing the immediate subcomponents, as they are
already processed by the upper loop.  Eventually we should get rid of
the iffy second loop, but I'm not sure how to, yet.

More importantly, this adds a regression test to make sure that when
we do refactor this code, the bug won't reappear.