Fix JSON conditional logic to work properly on global data (#14)
Before, conditional logic would only see the current row's data.  To
extend this, we need to pass on the full original data map to each
nested component.

Working on this exposed a potential issue: if "value" is used in the
form as a key, it will be clobbered by the current component in the
hierarchy.  This means conditional logic won't work properly if it
depends on this "value" element.

The proper fix would be to have the form property encode just the raw
JSON form and have another property for the parsed and Python-encoded
value.