Fix loading of tableComponents and correct unit tests
The unit tests were written like for dataGrid, but dataGrid is
"special" in that it can instantiate rows dynamically depending on the
input.  Table is more like the PanelComponent or such, which has a
predetermined set of rows and columns.  In each row/column, there will
be components with predetermined keys.

This also means we don't have to implement a getter for the rows, but
we can simply set the rows directly to the instantiated child
components when loading the data, as the structure is known by the
builder already.