Drop clumsy attempt at being generic from load_data
This really only caused more trouble than it helped - it turns out
that complex components require custom code anyway, so it makes more
sense to do that for these components and drop the buggy and difficult
to understand code which was a poor attempt at being generic.

This only broke the columnsComponent, so implement a proper custom
load_data there.  The fact that only this one broke points out that
this iffy code path was only really exercised by this component type
anyway.