Further simplification of standard load_data implementation
The remaining case for nested components was only exercised by
panelComponent, so move that bit there and drop it from the main
implementation (which now only works for standard input or display
components).

Doing so allows us to also get rid of the pretty awkward load_children
argument, which was introduced purely for addressComponent's sake,
which then necessarily infected the rest of the components as well.