Fix nested editGrid test targeting, fixup some other ids and count assertions
This to match the new structure of the nested form.

Also, the "editGrid1" component is not in the same row as dataGrid,
but on its own row!  Adjust to match.

And finally, component.datagrid is now component.grid, so that it
makes sense for both editgrid and datagrid.  Also adjust to match.