Improve selectComponent to support the Data Source Type: URL
Improve the selectComponent class to support the property Data
Source Type (dataSrc) with URL.
This applies to the methods value_label and value_labels, which
use the stored label(s).
So this doesn't execute a request to the URL, because we can't take
any control (e.g. due to authentication) over the host/server/service,